moved all the table descriptions to the knowledgebase and the project metaclass will manange the knowledgebase
